test(links): move qdrant-env-mutating controller tests to async false module
Both describe blocks mutated Application.put_env(:engram, :qdrant_url, ...)
in setup while the parent module was async: true, risking the race
described in docs/context/exunit-application-env-races.md. Moved to a
new async: false module.
